chore(executor): extract pg_on.rs (#116 step 9)
Pulls the PG-typed _on CRUD family out of mod.rs into a dedicated
file (181 LOC):

- insert_on / update_on / delete_on
- select_rows_on / select_one_row_on
- insert_returning_on
- bulk_insert_on

These accept any sqlx executor (&PgPool / &mut PgConnection /
Transaction). Tenant-scoped writes need them — schema-mode tenants
share the registry pool and rely on per-checkout SET search_path;
passing a fresh &PgPool would silently hit the wrong schema.

Mounted under #[cfg(feature = postgres)] in mod.rs, so per-function
postgres cfg gates were dropped inside the file. raw_query_on +
fetch_aggregate_on still live in mod.rs — they're at lines 1038+
in a different section and will move in a later step.

mod.rs: 2803 → 2637 LOC (-166). pg_on.rs: 181 LOC.

No public-API change. 1482 lib tests pass on sqlite,tenancy litmus.

Part of #116 (step 9).
