+++ crates/rustango/src/sql/executor/mod.rs
@@ -247,120 +247,17 @@ mod page;
 use page::inject_total_count;
 pub use page::Page;
 
-/// Like [`insert`] but accepts any sqlx executor — `&PgPool`,
-/// `&mut PgConnection`, or a transaction. Tenant-scoped writes need
-/// this: schema-mode tenants share the registry pool and rely on the
-/// per-checkout `SET search_path`, so passing `&PgPool` would silently
-/// hit the wrong schema.
-///
-/// # Errors
-/// As [`insert`].
-#[cfg(feature = "postgres")]
-pub async fn insert_on<'c, E>(executor: E, query: &InsertQuery) -> Result<(), ExecError>
-where
-    E: sqlx::Executor<'c, Database = sqlx::Postgres>,
-{
-    query.validate()?;
-    let stmt = Postgres.compile_insert(query)?;
-    let mut q: Query<'_, sqlx::Postgres, PgArguments> = sqlx::query(&stmt.sql);
-    for value in stmt.params {
-        q = bind_query(q, value);
-    }
-    q.execute(executor).await?;
-    Ok(())
-}
-
-/// Like [`insert_returning`] but accepts any sqlx executor.
-///
-/// # Errors
-/// As [`insert_returning`].
-#[cfg(feature = "postgres")]
-pub async fn insert_returning_on<'c, E>(
-    executor: E,
-    query: &InsertQuery,
-) -> Result<PgRow, ExecError>
-where
-    E: sqlx::Executor<'c, Database = sqlx::Postgres>,
-{
-    if query.returning.is_empty() {
-        return Err(ExecError::EmptyReturning);
-    }
-    query.validate()?;
-    let stmt = Postgres.compile_insert(query)?;
-    let mut q: Query<'_, sqlx::Postgres, PgArguments> = sqlx::query(&stmt.sql);
-    for value in stmt.params {
-        q = bind_query(q, value);
-    }
-    let row = q.fetch_one(executor).await?;
-    Ok(row)
-}
-
-/// Like [`bulk_insert`] but accepts any sqlx executor.
-///
-/// # Errors
-/// As [`bulk_insert`].
-#[cfg(feature = "postgres")]
-pub async fn bulk_insert_on<'c, E>(
-    executor: E,
-    query: &BulkInsertQuery,
-) -> Result<Vec<PgRow>, ExecError>
-where
-    E: sqlx::Executor<'c, Database = sqlx::Postgres>,
-{
-    query.validate()?;
-    let stmt = Postgres.compile_bulk_insert(query)?;
-    let mut q: Query<'_, sqlx::Postgres, PgArguments> = sqlx::query(&stmt.sql);
-    for value in stmt.params {
-        q = bind_query(q, value);
-    }
-    if query.returning.is_empty() {
-        q.execute(executor).await?;
-        Ok(Vec::new())
-    } else {
-        Ok(q.fetch_all(executor).await?)
-    }
-}
-
-/// Like [`update`] but accepts any sqlx executor.
-///
-/// # Errors
-/// As [`update`].
-#[cfg(feature = "postgres")]
-pub async fn update_on<'c, E>(executor: E, query: &UpdateQuery) -> Result<u64, ExecError>
-where
-    E: sqlx::Executor<'c, Database = sqlx::Postgres>,
-{
-    query.validate()?;
-    let stmt = Postgres.compile_update(query)?;
-    let mut q: Query<'_, sqlx::Postgres, PgArguments> = sqlx::query(&stmt.sql);
-    for value in stmt.params {
-        q = bind_query(q, value);
-    }
-    let result = q.execute(executor).await?;
-    Ok(result.rows_affected())
-}
-
-/// Like [`delete`] but accepts any sqlx executor.
-///
-/// # Errors
-/// As [`delete`].
-#[cfg(feature = "postgres")]
-pub async fn delete_on<'c, E>(executor: E, query: &DeleteQuery) -> Result<u64, ExecError>
-where
-    E: sqlx::Executor<'c, Database = sqlx::Postgres>,
-{
-    let stmt = Postgres.compile_delete(query)?;
-    let mut q: Query<'_, sqlx::Postgres, PgArguments> = sqlx::query(&stmt.sql);
-    for value in stmt.params {
-        q = bind_query(q, value);
-    }
-    let result = q.execute(executor).await?;
-    Ok(result.rows_affected())
-}
-
 // ====================================================================
-// row_to_json family — extracted to row_to_json.rs (#116 step 7)
+// PG `_on` CRUD family — extracted to pg_on.rs (#116 step 9)
 // ====================================================================
+
+#[cfg(feature = "postgres")]
+mod pg_on;
+#[cfg(feature = "postgres")]
+pub use pg_on::{
+    bulk_insert_on, delete_on, insert_on, insert_returning_on, select_one_row_on, select_rows_on,
+    update_on,
+};
 
 mod row_to_json;
 #[cfg(feature = "postgres")]
@@ -370,49 +267,6 @@ pub use row_to_json::row_to_json_my;
 #[cfg(feature = "sqlite")]
 pub use row_to_json::row_to_json_sqlite;
 pub use row_to_json::{select_one_row_as_json, select_rows_as_json};
-
-/// Like [`select_rows`] but accepts any sqlx executor — `&PgPool`,
-/// `&mut PgConnection`, or a `Transaction`. Required for tenancy
-/// projects whose per-request connection comes from
-/// [`crate::extractors::Tenant`] rather than a single global pool.
-///
-/// # Errors
-/// As [`select_rows`].
-#[cfg(feature = "postgres")]
-pub async fn select_rows_on<'c, E>(
-    executor: E,
-    query: &SelectQuery,
-) -> Result<Vec<PgRow>, ExecError>
-where
-    E: sqlx::Executor<'c, Database = sqlx::Postgres>,
-{
-    let stmt = Postgres.compile_select(query)?;
-    let mut q: Query<'_, sqlx::Postgres, PgArguments> = sqlx::query(&stmt.sql);
-    for value in stmt.params {
-        q = bind_query(q, value);
-    }
-    Ok(q.fetch_all(executor).await?)
-}
-
-/// Like [`select_one_row`] but accepts any sqlx executor.
-///
-/// # Errors
-/// As [`select_one_row`].
-#[cfg(feature = "postgres")]
-pub async fn select_one_row_on<'c, E>(
-    executor: E,
-    query: &SelectQuery,
-) -> Result<Option<PgRow>, ExecError>
-where
-    E: sqlx::Executor<'c, Database = sqlx::Postgres>,
-{
-    let stmt = Postgres.compile_select(query)?;
-    let mut q: Query<'_, sqlx::Postgres, PgArguments> = sqlx::query(&stmt.sql);
-    for value in stmt.params {
-        q = bind_query(q, value);
-    }
-    Ok(q.fetch_optional(executor).await?)
-}
 
 /// Slice 9.0b — annotate each parent row with the COUNT of its
 /// children, returning `Vec<(Parent, i64)>` from a **single** SQL:
